pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=5.0.0",
    "fastapi>=0.118.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.11.10",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
    "websockets>=15.0.1",
]
